requirements.txt
python-telegram-bot==20.5
pandas==2.0.3
tabulate==0.9.0
openpyxl==3.1.2
python-dotenv==1.0.0